% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{aug_genlassoRcpp}
\alias{aug_genlassoRcpp}
\title{Solving Generalized LASSO with fixed \eqn{\lambda = 1}}
\usage{
aug_genlassoRcpp(
  y,
  W,
  m,
  c,
  lambda1,
  lambda2,
  global_rho,
  a,
  rho,
  max_iter,
  eps
)
}
\arguments{
\item{reports}{A binary matrix. Each row is a report}

\item{n_drugs}{The number of drugs}

\item{n_events}{The number of events}
}
\value{
The estimated vector \eqn{\hat{\beta}}
}
\description{
Solves efficiently the generalized LASSO problem of the form 
\deqn{
  \hat{\beta} = \text{argmin } \frac{1}{2} || y - \beta ||_2^2 + ||D\beta||_1 
}
where \eqn{\beta} and \eqn{y} are \eqn{m}-dimensional vectors and 
\eqn{D} is a \eqn{(c \times m)}-matrix where \eqn{c \geq m}. 
We solve this optimization problem using an adaption of the ADMM
algorithm presented in Zhu (2017).
}
\references{
Zhu, Y. (2017). An Augmented ADMM Algorithm With Application to the 
Generalized Lasso Problem. Journal of Computational and Graphical Statistics, 
26(1), 195–204. https://doi.org/10.1080/10618600.2015.1114491
}
\seealso{
\code{\link{convertRawReports2Tables}}
}
