% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-correctness-input.R
\name{check_correctness_input}
\alias{check_correctness_input}
\title{Check whether Input is Valid}
\usage{
check_correctness_input(raw_data, W, lambda1, lambda2, rho)
}
\arguments{
\item{raw_data}{A list with matrices. The number of columns should be the 
same for each matrix}

\item{W}{The \eqn{(m \times m)}-dimensional upper-triangular 
weight matrix \eqn{W}}

\item{lambda1}{The \eqn{\lambda_1} LASSO penalty term}

\item{lambda2}{The \eqn{\lambda_2} global smoothing parameter}

\item{rho}{The \eqn{\rho} ADMM's penalty parameter}
}
\description{
Checks whether the input for the function \code{\link{CVN}} is valid.
This function does not return anything. The execution of the function
halts when an issue has been detected.
}
\seealso{
\code{\link{CVN}}
}
