% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine-cvn.R
\name{combine_cvn}
\alias{combine_cvn}
\title{Combine CVNs of size 1}
\usage{
combine_cvn(cvn_list, minimal = TRUE)
}
\arguments{
\item{cvn_list}{a list of CVNs, each of size 1 (one lambda/gamma pair)}

\item{minimal}{logical; stripped CVN (default = TRUE)}
}
\value{
a CVN of size length(cvn_list)
}
\description{
Function to combine single CVNs. Handy, if CVNs are computed in parallel using
foreach loop and n_cores = 1
}
\examples{
data(grid)

#' Choice of the weight matrix W. Each of 2 covariates has 3 categories
#' (uniform random)
W_grid <- create_weight_matrix("grid", k = 3, l = 3)

fit1 <- CVN(data = grid, W = W_grid, lambda1 = 1, lambda2 = 0.5, 
            gamma_ebic = 0.75,
            n_cores = 1, eps = 1e-2, maxiter = 200, minimal = TRUE)
            
fit2 <- CVN(data = grid, W = W_grid, lambda1 = 2, lambda2 = 0.75, 
            gamma_ebic = 0.75,
            n_cores = 1, eps = 1e-2, maxiter = 200, minimal = TRUE)
            
(fit <- combine_cvn(list(fit1, fit2)))            
}
