% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-connList-flsa.R
\name{create_connListObj}
\alias{create_connListObj}
\title{Create a connListObj for FLSA package}
\usage{
create_connListObj(W)
}
\arguments{
\item{W}{A weight matrix.}
}
\value{
A connListObj for FLSA package and a logical value indicating whether 
        all elements of connList are NULL.
}
\description{
This function creates a connListObj for the FLSA package given a weight matrix.
}
