% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-CVN-data-object.R
\name{create_CVN_data_object}
\alias{create_CVN_data_object}
\title{Create a CVN Data Object}
\usage{
create_CVN_data_object(raw_data, W, normalized = TRUE, verbose = TRUE)
}
\arguments{
\item{raw_data}{A list with matrices. The number of columns should be the 
same for each matrix}

\item{W}{The \eqn{(m x m)}-dimensional uppertriangular weight matrix}

\item{normalized}{Data is normalized if TRUE. Otherwise the data is only
centered (Default: TRUE)}

\item{Verbose}{Verbose (Default: FALSE)}
}
\value{
A CVN data object; a list with entries
  \item{\code{data}}{The \code{raw_data}, but then normalized or centered}
  \item{\code{W}}{The \eqn{m x m} weight matrix}
  \item{\code{m}}{Number of graphs}
  \item{\code{p}}{Number of variables}
  \item{\code{n_obs}}{Number of observations for each graph}
}
\description{
Creates a Covariate-varying network (CVN) object for estimating the 
graphs. It only requires the \code{raw_data}: a list of \code{m} different
matrices, each containing the observed data. Each column is a variable. Therefore,
each matrix should have the same number of columns. The number of observations
can differ between the datasets.
}
