% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-weight-matrices.R
\name{create_grid_adjacency}
\alias{create_grid_adjacency}
\title{Create grid adjacency}
\usage{
create_grid_adjacency(k, l)
}
\arguments{
\item{k}{Number of categories in the first external covariate}

\item{l}{Number of categories in the second external covariate}
}
\value{
A matrix
}
\description{
An algorithm which creates a weight matrix in grid structure
}
\examples{
CVN:::create_grid_adjacency(k = 2, l = 3)
}
\keyword{internal}
