% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-matrix-D.R
\name{create_matrix_D}
\alias{create_matrix_D}
\title{Create matrix D to be used for the Generalized LASSO}
\usage{
create_matrix_D(W, lambda1, lambda2, rho = 1)
}
\arguments{
\item{W}{The \code{m x m}-dimensional uppertriangular weight matrix}

\item{lambda1}{The \eqn{\lambda1} LASSO penalty term}

\item{lambda2}{The \eqn{\lambda2} global smoothing parameter}

\item{rho}{The \eqn{\rho} ADMM's penalty parameter (Default: 1)}
}
\value{
A \eqn{((m*(m+1)/2) x m)}-matrix
}
\description{
Generates a matrix D to be used for the generalized LASSO. 
We solve a generalized LASSO problem for each egde \eqn{(s,t)} 
for each update step for \eqn{Z(k+1)}.
}
\references{
Tibshirani, R. J., & Taylor, J. (2011). 
            The solution path of the generalized lasso. 
            Annals of Statistics, 39(3), 1335–1371. 
            https://doi.org/10.1214/11-AOS878
}
