% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-graph.R
\name{create_nodes_visnetwork}
\alias{create_nodes_visnetwork}
\title{Nodes for the \code{visNetwork} package}
\usage{
create_nodes_visnetwork(n_nodes, labels = 1:n_nodes)
}
\arguments{
\item{n_nodes}{Number of nodes in the graph}

\item{labels}{The labels for the individual nodes
(Default: \code{1:n_nodes})}
}
\value{
Data frame with two columns: \code{id} and \code{title}
}
\description{
Creates a data frame that can be used for the
\code{visNetwork} package.
}
