% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-weight-matrices.R
\name{create_weight_matrix}
\alias{create_weight_matrix}
\title{Different Weight Matrices}
\usage{
create_weight_matrix(
  type = c("full", "glasso", "grid", "uniform-random"),
  k,
  l,
  plot = TRUE
)
}
\arguments{
\item{type}{The type of weight matrix}

\item{k}{Number of categories in the first external covariate}

\item{l}{Number of categories in the second external covariate}

\item{plot}{If TRUE, the weight matrix is plotted (default: TRUE)}
}
\value{
Weight matrix
}
\description{
This function generates different weight matrices for two external covariates, \eqn{W}. 
There are several types: 
\describe{
\item{\code{full} }{All graphs are fully connected with weight 1}
\item{\code{glasso} }{All graphs are disconnected with weight 0. This 
             mimicks the GLASSO, where each graph is estimated independently}
\item{\code{grid} }{A weight matrix for a \eqn{k \times l} grid}
\item{\code{uniform-random} }{Fully-connected, but the entries are drawn 
                   from a uniform distribution}
}
}
\examples{
W <- create_weight_matrix(type="grid", k=3, l=2)

}
