% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVN-package.R
\docType{package}
\name{CVN-package}
\alias{CVN-package}
\title{CVN: Covariate-varying Networks}
\description{
A package for inferring high-dimensional Gaussian graphical networks that change with multiple discrete covariates
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bips-hb/CVN}
  \item Report bugs at \url{https://github.com/bips-hb/CVN/issues}
}

}
\keyword{internal}
