% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVN.R
\name{CVN}
\alias{CVN}
\title{Inferring a Covariate-Varying Network (CVN)}
\usage{
CVN(
  data,
  W,
  lambda1 = 1,
  lambda2 = 1,
  rho = 1,
  epsilon = 10^(-5),
  maxiter = 100,
  n_cores = 1,
  truncate = 1e-05,
  normalized = FALSE,
  warmstart = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{data}{A list with matrices. The number of columns should be the 
same for each matrix. Number of observations can differ}

\item{W}{The \code{(m x m)}-dimensional uppertriangular weight matrix}

\item{lambda1}{The \eqn{\lambda1} LASSO penalty term}

\item{lambda2}{The \eqn{\lambda2} global smoothing parameter}

\item{rho}{The \eqn{\rho} ADMM's penalty parameter (Default = 1)}

\item{epsilon}{If the relative difference between two update steps is 
smaller than \eqn{\epsilon}, the algorithm stops
(Default: \code{1e-5})}

\item{maxiter}{Maximum number of iterations (Default = 100)}

\item{n_cores}{Number of cores used (Default = 1)}

\item{truncate}{All values of the final Theta below \code{truncate} will be 
set to \code{0}. (Default = \code{1e-5})}

\item{normalized}{Data is normalized if \code{TRUE}. Otherwise the data is only
centered (Default: \code{FALSE})}

\item{warmstart}{If \code{TRUE}, use the \code{huge} package for estimating
the individual graphs first (Default: \code{FALSE})}

\item{verbose}{Verbose (Default: \code{FALSE})}
}
\value{
A CVN object; a list with entries
   \item{\code{Theta}}{The estimated precision matrices}
   \item{\code{adj_matrices}}{The estimated adjacency matrices; 1 if there is an edge, 0 otherwise}
   \item{\code{Sigma}}{Empirical covariance matrices}
   \item{\code{m}}{Number of graphs}
   \item{\code{p}}{Number of variables}
   \item{\code{n_obs}}{Vector of length \eqn{m} with number of observations for each graph}
  \item{\code{data}}{The \code{data}, but then normalized or centered}
  \item{\code{normalized}}{If \code{TRUE}, \code{data} was normalized. Otherwise \code{data} was only centered}
  \item{\code{W}}{The \code{m x m} weight matrix}
  \item{\code{D}}{Matrix \eqn{D} used for the Generalized LASSO}
  \item{\code{lambda1}}{The \eqn{\lambda1} LASSO penalty term}
  \item{\code{lambda2}}{The \eqn{\lambda2} global smoothing parameter} 
  \item{\code{rho}}{The \eqn{\rho} ADMM's penalty parameter} 
  \item{\code{epsilon}}{The stopping criterion \eqn{\epsilon}} 
  \item{\code{converged}}{If \code{TRUE}, stopping condition has been met}
  \item{\code{value}}{The final relative difference}
  \item{\code{n_iterations}}{Number of iterations}
  \item{\code{truncate}}{Truncation value for \eqn{\Theta}}
}
\description{
Estimates a covariate-varying network model (CVN), i.e., \eqn{m} 
Gaussian graphical models that change with (multiple) external covariate(s). 
The smoothing between the graphs is specified by the \eqn{(m x m)}-dimensional
weight matrix \eqn{W}. The function returns the estimated precision matrices 
for each graph.
}
