% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine-information-criterion.R
\name{determine_information_criterion_cvn}
\alias{determine_information_criterion_cvn}
\title{Information Criteria for a \code{cvn} object}
\usage{
determine_information_criterion_cvn(cvn, gamma = 0.5)
}
\arguments{
\item{cvn}{A CVN object, see \code{\link{CVN}}}

\item{gamma}{The gamma value for the eBIC (default: 0.5)}
}
\description{
Determines a given information criteria for a \code{cvn} object,
see \code{\link{CVN}}.
}
\examples{
path <- system.file("cvnfit.rda", package = "CVN")
load(path)
determine_information_criterion_cvn(fit, gamma = 0.7)
}
