% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate}
\alias{estimate}
\title{Estimate a CVN}
\usage{
estimate(
  m,
  p,
  W,
  Theta0,
  Z0,
  Y0,
  a,
  eta1,
  eta2,
  Sigma,
  n_obs,
  rho,
  rho_genlasso,
  eps,
  eps_genlasso,
  maxiter,
  maxiter_genlasso,
  truncate,
  truncate_genlasso,
  verbose = FALSE
)
}
\description{
A function for estimating a CVN for a single \eqn{(\lambda_1, \lambda_2)}-value. 
See for more details \code{\link{CVN}}
}
\keyword{internal}
