% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_cvn.R
\name{extract_cvn}
\alias{extract_cvn}
\title{Extract CVN}
\usage{
extract_cvn(cvn, id)
}
\arguments{
\item{cvn}{A CVN object, see \code{\link{CVN}}}

\item{id}{Integer; Which id from the CVN object should be extracted}
}
\value{
A CVN object
}
\description{
Function that extracts one CVN if the CVN object contains more than one. 
Its helpful when just one tuning parameter constellation wants to be, e.g., plotted.
}
\examples{
\dontrun{
# Example code of usage (not run because of longer running time)

data(grid)
W <- create_weight_matrix("grid", 3, 3)
fit <-  CVN(grid, W, n_cores = 1,
            lambda1 = c(0.5, 1), 
            lambda2 = c(0.1, 0.5)) 
(fit2 <- extract_cvn(fit, id = 2))
}


}
