% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-core-graph.R
\name{find_core_graph}
\alias{find_core_graph}
\title{The Core Graph}
\usage{
find_core_graph(cvn)
}
\arguments{
\item{cvn}{A \code{cvn} object}
}
\value{
A list of adjacency matrix, one for each 
        value of (lambda1, lambda2)
}
\description{
Finds the 'core graph', i.e., those edges that are 
present in all of the estimated graphs
}
\examples{
path <- system.file("cvnfit.RData", package = "CVN")
load(path)
find_core_graph(fit) 

}
