% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-core-graph.R
\name{find_unique_edges}
\alias{find_unique_edges}
\title{Unique Edges}
\usage{
find_unique_edges(cvn)
}
\arguments{
\item{cvn}{A \code{cvn} object}
}
\value{
A list of adjacency matrix, one for each 
        value of (lambda1, lambda2), showing the edges that are unique for each
        single subgraph.
}
\description{
Function that finds edges that are only present in one subgraph of the CVN model.
}
\examples{
data(grid)

W <- create_weight_matrix("grid", 3, 3)

# lambdas:
lambda1 = 1  # can also be lambda1 = 1:2 
lambda2 = .2

fit <- CVN(grid, W, lambda1 = lambda1, lambda2 = lambda2)
                 
find_unique_edges(fit)
}
