% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genlasso-wrapper.R
\name{genlasso_wrapper}
\alias{genlasso_wrapper}
\title{Wrapper for \code{genlassoRcpp}}
\usage{
genlasso_wrapper(y, W, m, eta1, eta2, a, rho, max_iter, eps, truncate)
}
\arguments{
\item{y}{A numeric vector}

\item{W}{Weight matrix}

\item{m}{Number of graphs}

\item{eta1}{lambda1 / rho, with \eqn{\rho} is penalty parameter for the global ADMM algorithm (Default: \code{1})}

\item{eta2}{lambda2 / rho}

\item{a}{constant}

\item{rho}{the \eqn{\rho} penalty parameter for the ADMM algorithm}

\item{max_iter}{Maximum number of iterations (Default: \code{100})}

\item{eps}{If the relative difference between two update steps is 
smaller than \eqn{\epsilon}, the algorithm stops. 
See \code{\link[CVN]{relative_difference_precision_matrices}}
(Default: \code{1e-4})}

\item{truncate}{All values of the final \eqn{\hat{\Theta}_i}'s below \code{truncate} will be 
set to \code{0}. (Default: \code{1e-4})}
}
\description{
See for details \code{\link{genlassoRcpp}}
}
\seealso{
\code{\link[CVN]{relative_difference_precision_matrices}}

\code{\link{genlassoRcpp}}
}
\keyword{internal}
