% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{grid}
\alias{grid}
\title{Data for a grid of graphs (3 x 3)}
\format{
A list of length 9
}
\usage{
data(grid)
}
\description{
Data generated for 9 graphs in total, organized in a grid of 
(3x3). See the package \code{CVNSim} for more information 
on how the grid is constructed: \url{https://github.com/bips-hb/CVNSim}
}
\references{
\url{https://github.com/bips-hb/CVNSim}
}
\keyword{datasets}
