% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamming-distance.R
\name{hamming_distance_adj_matrices}
\alias{hamming_distance_adj_matrices}
\title{Structural Hamming Distance}
\usage{
hamming_distance_adj_matrices(adj_matrix1, adj_matrix2)
}
\arguments{
\item{adj_matrix1, adj_matrix2}{Two adjacency matrices}
}
\value{
Hamming distance
}
\description{
Returns the structural Hamming distance between two
graphs
}
