% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hd-weight-matrix.R
\name{hd_weight_matrix}
\alias{hd_weight_matrix}
\title{Heatmap Plot of the Weight Matrix}
\usage{
hd_weight_matrix(
  W,
  title = "",
  legend_label = "weight",
  add_counts_to_cells = TRUE,
  add_ticks_labels = TRUE,
  t = -6,
  r = -8
)
}
\arguments{
\item{W}{Weight matrix}

\item{title}{Title plot (Default is none)}

\item{legend_label}{Title of the legend (Default: "weight")}

\item{add_counts_to_cells}{If \code{TRUE}, counts from the matrix are
added to the plot (Default: \code{TRUE})}

\item{add_ticks_labels}{If \code{TRUE}, the number corresponding to the graph
is add to the plot (Default: \code{TRUE})}

\item{t}{Distance between tick labels and x-axis (Default: -6)}

\item{r}{Distance between tick labels and y-axis (Default: -8)}
}
\value{
A heatmap plot
}
\description{
Returns a heatmap of the weight matrix
}
\examples{
W_uniform <- round(create_weight_matrix(type="uniform-random", 3, 2), 2)
hd_weight_matrix(W_uniform, title = "2 external covariates with 3 and 2 categories respectively")
            
}
\seealso{
\code{\link{plot_weight_matrix}}
}
