% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hits-end-lambda-intervals.R
\name{hits_end_lambda_intervals}
\alias{hits_end_lambda_intervals}
\title{Hitting the end points of \eqn{(\lambda_1, \lambda_2)} interval}
\usage{
hits_end_lambda_intervals(results)
}
\arguments{
\item{results}{Results of the \code{\link{CVN}} function}
}
\value{
List with two values:
   \item{\code{hits_border_aic}}{If \code{TRUE}, hits the border for the AIC}
   \item{\code{hits_border_bic}}{If \code{TRUE}, hits the border for the BIC}
}
\description{
One often selected the optimal model for the \eqn{(\lambda_1, \lambda_2)}-values
based on the AIC and BIC.
This function checks and warns when the optimal value lies on the border of the
values \eqn{(\lambda_1, \lambda_2)} takes.
}
\keyword{internal}
