% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is-binary-matrix.R
\name{is_binary_matrix}
\alias{is_binary_matrix}
\title{Check if a matrix is binary}
\usage{
is_binary_matrix(mat)
}
\arguments{
\item{mat}{A matrix to be tested.}
}
\value{
A logical value indicating whether the matrix is binary or not.
}
\description{
This function checks whether a given matrix is binary, i.e., all its elements are either 0 or 1.
}
\examples{
binary_mat <- matrix(c(0, 1, 1, 0, 1, 0, 0, 1), nrow = 2)
is_binary_matrix(binary_mat)

}
