% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.cvn.R
\name{plot.cvn}
\alias{plot.cvn}
\title{Plot Function for CVN Object Class}
\usage{
\method{plot}{cvn}(x, ...)
}
\arguments{
\item{x}{\code{cvn} object}

\item{...}{Additional arguments to pass to \code{\link{CVN}}}
}
\description{
Custom plot method for CVN objects.
}
\seealso{
\code{CVN}
}
