% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-aic.R
\name{plot_aic}
\alias{plot_aic}
\title{Heat Map of Performance (AIC)}
\usage{
plot_aic(
  cvn,
  limits = c(NA, NA),
  title = "",
  xlabel = "lambda1",
  ylabel = "lambda2",
  legend_label = "AIC"
)
}
\arguments{
\item{cvn}{}

\item{limits}{The limits for the values of the Hamming distance}

\item{title}{Title plot (Default is none)}
}
\value{
A heatmap plot
}
\description{
Returns a heat map of the AIC
}
