% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-distances.R
\name{plot_hamming_distances}
\alias{plot_hamming_distances}
\title{Heat Map of the Distances between Graphs}
\usage{
plot_hamming_distances(
  distance_matrix,
  absolute = TRUE,
  limits = c(NA, NA),
  title = "",
  legend_label = "Hamming Distance",
  add_counts_to_cells = TRUE,
  add_ticks_labels = TRUE,
  t = -6,
  r = -8
)
}
\arguments{
\item{distance_matrix}{Symmetric matrix with distances}

\item{absolute}{If \code{FALSE}, rescaled to \verb{[0,1]}}

\item{limits}{The limits for the values of the Hamming distance}

\item{title}{Title plot (Default is none)}

\item{legend_label}{Title of the legend (Default: "Hamming Distance")}

\item{add_counts_to_cells}{If \code{TRUE}, counts from the matrix are
added to the plot (Default: \code{TRUE})}

\item{add_ticks_labels}{If \code{TRUE}, the number corresponding to the graph
is add to the plot (Default: \code{TRUE})}

\item{t}{Distance between tick labels and x-axis (Default: -6)}

\item{r}{Distance between tick labels and y-axis (Default: -8)}
}
\value{
A heatmap plot
}
\description{
Returns a heat map of the distance matrix for
a particular CVN
}
\examples{
         
path <- system.file("cvnfit.rda", package = "CVN")
load(path)
hd <- hamming_distance(fit, verbose = FALSE)

plot_hamming_distances(hd$distance[[1]])
}
