% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-distances.R
\name{plot_hamming_distances_cvn}
\alias{plot_hamming_distances_cvn}
\title{Heatmaps for a CVN}
\usage{
plot_hamming_distances_cvn(
  cvn,
  absolute = TRUE,
  same_range = TRUE,
  titles = rep("", cvn$n_lambda_values),
  legend_label = "Hamming Distance",
  add_counts_to_cells = TRUE,
  add_ticks_labels = TRUE,
  t = -6,
  r = -8,
  verbose = TRUE
)
}
\arguments{
\item{cvn}{A \code{cvn} object}

\item{absolute}{If \code{FALSE}, rescaled to \verb{[0,1]}}

\item{same_range}{If \code{TRUE}, all heatmaps have the same range of values
of the Hamming distance shown (Default: TRUE)}

\item{titles}{Title of the plots (Default is none)}

\item{legend_label}{Title of the legend (Default: "Hamming Distance")}

\item{add_counts_to_cells}{If \code{TRUE}, counts from the matrix are
added to the plot (Default: \code{TRUE})}

\item{add_ticks_labels}{If \code{TRUE}, the number corresponding to the graph
is add to the plot (Default: \code{TRUE})}

\item{t}{Distance between tick labels and x-axis (Default: -6)}

\item{r}{Distance between tick labels and y-axis (Default: -8)}

\item{verbose}{If \code{TRUE}, shows progress bar (Default: \code{TRUE})}
}
\value{
List of plots
}
\description{
Creates all the heatmaps for a CVN, a heatmap for each
pair of \eqn{(\lambda_1, \lambda_2)}
}
\examples{
path <- system.file("cvnfit.rda", package = "CVN")
load(path)

plot_hamming_distances_cvn(fit)

}
