% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-information-criterion.R
\name{plot_information_criterion}
\alias{plot_information_criterion}
\title{Heat Map of an Information Criterion (AIC or BIC)}
\usage{
plot_information_criterion(
  cvn,
  criterion = c("bic", "aic", "ebic"),
  use_gammas = TRUE,
  show_minimum = TRUE,
  title = "",
  xlabel = NULL,
  ylabel = NULL,
  legend_label = NULL,
  limits = c(NA, NA)
)
}
\arguments{
\item{cvn}{Fitted CVN, see \code{\link{CVN}}}

\item{criterion}{The information criterion, must be either \code{'aic'} 
or \code{'bic'}. Default: \code{'bic'}}

\item{use_gammas}{If \code{TRUE}, plots the \eqn{\gamma}-values. Otherwise, 
the \eqn{\lambda}-values are used}

\item{show_minimum}{If \code{TRUE}, an orange dot is put on the point with the 
minimum value of the information criterion is. If \code{FALSE}, 
no dot is added. Default: \code{TRUE}.}

\item{title}{Title plot (Default is none)}

\item{xlabel}{Label for the \eqn{x}-axis. Default depends on \code{use_gammas}. 
If \code{use_gammas = TRUE}, then the label is 'gamma1'. Otherwise, 
'lambda1'}

\item{ylabel}{Label for the \eqn{x}-axis. Default depends on \code{use_gammas}. 
If \code{use_gammas = TRUE}, then the label is 'gamma1'. Otherwise, 
'lambda1'}

\item{legend_label}{Title for the legend. Default depends on \code{criterion}. 
If \code{'aic'}, then the label is 'AIC'. 
Otherwise, 'BIC'.}

\item{limits}{The limits for the values of the Hamming distance}
}
\value{
A heatmap plot
}
\description{
Returns a heat map of the AIC or BIC for a fitted CVN
}
\examples{
data(grid)
m <- 9

W <- create_weight_matrix(type="grid", 3, 3, plot=FALSE)
(cvn <- CVN::CVN(grid, W, lambda1 = 2, lambda2 = .5))

plot_information_criterion(cvn, criterion = "aic")
plot_information_criterion(cvn, criterion = "bic")
plot_information_criterion(cvn, criterion = "ebic")

}
