% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-weight-matrix.R
\name{plot_weight_matrix}
\alias{plot_weight_matrix}
\title{Plot Weight Matrix}
\usage{
plot_weight_matrix(
  W,
  title = "",
  legend_label = "weight",
  add_counts_to_cells = TRUE,
  add_ticks_labels = TRUE,
  t = -6,
  r = -8
)
}
\arguments{
\item{W}{Symmetric weight matrix}

\item{title}{Title plot (Default is none)}

\item{legend_label}{Title of the legend (Default: "weight")}

\item{add_counts_to_cells}{If \code{TRUE}, counts from the matrix are 
added to the plot (Default: \code{TRUE})}

\item{add_ticks_labels}{If \code{TRUE}, the number corresponding to the graph
is add to the plot (Default: \code{TRUE})}

\item{t}{Distance between tick labels and x-axis (Default: -6)}

\item{r}{Distance between tick labels and y-axis (Default: -8)}
}
\value{
A heatmap plot
}
\description{
Returns a heat map of a weight matrix
}
