% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relative-difference-precision-matrices.R
\name{relative_difference_precision_matrices}
\alias{relative_difference_precision_matrices}
\title{The Relative Difference between two Precision Matrices}
\usage{
relative_difference_precision_matrices(Theta_new, Theta_old)
}
\arguments{
\item{Theta_new}{A list with matrices with the updated values of \eqn{\Theta}}

\item{Theta_old}{A list with matrices with the old values of \eqn{\Theta}}
}
\value{
The relative difference between \eqn{\Theta(k+1)} and \eqn{\Theta(k)}
}
\description{
Returns the relative \eqn{L1} difference between precision matrix \eqn{\Theta(k+1)}
(parameter \code{Theta_new}) and \eqn{\Theta(k)} (parameter \code{Theta_old}).
}
\details{
This is used for checking whether the stopping condition has been met.
}
