% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-Theta.R
\name{updateTheta}
\alias{updateTheta}
\title{The \eqn{\Theta}-update step for the ADMM}
\usage{
updateTheta(Theta, Z, Y)
}
\arguments{
\item{Theta}{A list with matrices with the current values of \eqn{\Theta}}

\item{Z}{A list with matrices with the current values of \eqn{Z}}

\item{Y}{A list with matrices with the current values of \eqn{Y}}
}
\value{
A list with matrices with the new values of \eqn{\Theta}
}
\description{
Returns the updated value of \eqn{\Theta} for the ADMM given the previously 
updated values of \eqn{Z} and \eqn{Y}
}
\keyword{internal}
