% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-Y.R
\name{updateY}
\alias{updateY}
\title{The \eqn{Y}-update step for the ADMM}
\usage{
updateY(Theta, Z, Y)
}
\arguments{
\item{Theta}{A list with matrices with the current values of \eqn{\Theta}}

\item{Z}{A list with matrices with the current values of \eqn{Z}}

\item{Y}{A list with matrices with the current values of \eqn{Y}}
}
\value{
A list with matrices with the new values of \eqn{Y}
}
\description{
Returns the updated value of \eqn{Y} for the ADMM given the previously 
updated values of \eqn{\Theta} and \eqn{Z}
}
\keyword{internal}
