% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-Z.R
\name{updateZ}
\alias{updateZ}
\title{The \eqn{Z}-update step for the ADMM}
\usage{
updateZ(Theta, Z, Y)
}
\arguments{
\item{Theta}{A list with matrices with the current values of \eqn{\Theta}}

\item{Z}{A list with matrices with the current values of \eqn{Z}}

\item{Y}{A list with matrices with the current values of \eqn{Y}}
}
\value{
A list with matrices with the new values of \eqn{Z}
}
\description{
Returns the updated value of \eqn{Z} for the ADMM given the previously 
updated values of \eqn{\Theta} and \eqn{Y}
}
