% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-Z-flsa-package.R
\name{updateZ_flsa_package}
\alias{updateZ_flsa_package}
\title{Update Z using FLSA package}
\usage{
updateZ_flsa_package(m, p, Theta, Y, connListObj, connList_is_null, eta1, eta2)
}
\arguments{
\item{m}{Number of graphs}

\item{p}{Number of covariates}

\item{Theta}{List of matrices}

\item{Y}{List of matrices}

\item{connListObj}{connList object for FLSA package}

\item{connList_is_null}{Logical value indicating whether all elements of connList are NULL}

\item{eta1}{Sparsity parameter}

\item{eta2}{Smoothness parameter}
}
\value{
Updated Z matrix
}
\description{
This function updates Z matrix using FLSA package.
}
