% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-Z-wrapper.R
\name{updateZ_wrapper}
\alias{updateZ_wrapper}
\title{Wrapper for update-Z}
\usage{
updateZ_wrapper(
  m,
  p,
  nrow_D,
  Theta,
  Y,
  W,
  eta1,
  eta2,
  a,
  rho_genlasso,
  maxiter_genlasso,
  eps_genlasso,
  truncate_genlasso
)
}
\description{
Wrapper for the updateZRcpp function
See \code{\link{updateZ}} for more 
information
}
