% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{updateZRcpp}
\alias{updateZRcpp}
\title{Solving Generalized LASSO with fixed \eqn{\lambda = 1}}
\usage{
updateZRcpp(m, p, c, Theta, Y, W, eta1, eta2, a, rho, max_iter, eps, truncate)
}
\arguments{
\item{m}{The number of graphs}

\item{c}{Number of rows of matrix \eqn{D}, which is equal to 
\eqn{c = m + (m(m-1))/2}}

\item{W}{The weight matrix \eqn{W} of dimensions \eqn{m x m}}

\item{eta1}{Equals \eqn{\lambda_1 / rho}}

\item{eta2}{Equals \eqn{\lambda_2 / rho}}

\item{a}{Value added to the diagonal of \eqn{-D'D} so that
the matrix is positive definite, see 
\code{\link{matrix_A_inner_ADMM}}}

\item{rho}{The ADMM's parameter}

\item{max_iter}{Maximum number of iterations}

\item{eps}{Stopping criterion. If differences 
are smaller than \eqn{\epsilon}, algorithm
is halted}

\item{truncate}{Values below \code{truncate} are 
set to \code{0}}

\item{y}{The \eqn{y} vector of length \eqn{m}}
}
\value{
The estimated vector \eqn{\hat{\beta}}
}
\description{
Solves efficiently the generalized LASSO problem of the form 
\deqn{
  \hat{\beta} = \text{argmin } \frac{1}{2} || y - \beta ||_2^2 + ||D\beta||_1 
}
where \eqn{\beta} and \eqn{y} are \eqn{m}-dimensional vectors and 
\eqn{D} is a \eqn{(c \times m)}-matrix where \eqn{c \geq m}. 
We solve this optimization problem using an adaption of the ADMM
algorithm presented in Zhu (2017).
}
\seealso{
\code{\link{updateZ_wrapper}}
}
