% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-graph.R
\name{visnetwork_cvn}
\alias{visnetwork_cvn}
\title{All \code{visNetwork} plots for a CVN object}
\usage{
visnetwork_cvn(
  cvn,
  node_titles = 1:cvn$p,
  titles = lapply(1:cvn$n_lambda_values, function(i) sapply(1:cvn$m, function(j) "")),
  show_core_graph = TRUE,
  width = c(3, 1),
  color = c("red", "blue"),
  igraph_layout = "layout_in_circle"
)
}
\description{
Creates all \code{visNetwork} plots, see \code{\link{CVN::visnetwork}}, 
for all graphs in a \code{cvn} object
}
