% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{check_correctness_input}
\alias{check_correctness_input}
\title{Check whether the Input is Valid}
\usage{
check_correctness_input(raw_data, W, lambda1, lambda2, rho)
}
\arguments{
\item{raw_data}{A list with matrices. The number of columns should be the 
same for each matrix}

\item{W}{The \eqn{(m x m)}-dimensional uppertriangular weight matrix}

\item{lambda1}{LASSO penalization parameter \eqn{\lambda_1}}

\item{lambda2}{Global smoothing parameter \eqn{\lambda2}}

\item{rho}{ADMM's penalty parameter \eqn{\rho}}
}
\description{
Checks whether the input for the function \code{CVN} is valid.
This function does not return anything. The execution of the function
halts when an issue has been detected
}
\seealso{
\code{\link{CVN}}
}
