% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_cvn_interpolated.R
\name{combine_cvn_interpolated}
\alias{combine_cvn_interpolated}
\title{Combine CVN object with interpolated CVN object}
\usage{
combine_cvn_interpolated(cvn, cvn_interpolated)
}
\arguments{
\item{cvn}{A CVN fit with \eqn{m} graphs}

\item{cvn_interpolated}{A interpolated CVN from a fitted CVN model with \eqn{m} graphs}
}
\value{
A 'cvn_interpolated' object (see \link{\code{interpolate}}).
}
\description{
Combine an interpolated CVN to an original CVN
}
