% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-graph.R
\name{create_edges_visnetwork}
\alias{create_edges_visnetwork}
\title{Create a \code{data.frame} for the Edges for \code{visNetwork}}
\usage{
create_edges_visnetwork(adj_matrix)
}
\arguments{
\item{adj_matrix}{A symmetric adjacency matrix}
}
\value{
Data frame that be used as input for \code{visNetwork}
}
\description{
In order to visualize a graph, we need to create a
\code{data.frame} that can be used by the \code{visNetwork} package.
This function returns the needed \code{data.frame} given a
adjacency matrix.
}
\examples{
nodes <- create_nodes_visnetwork(n_nodes = 5, labels = LETTERS[1:5])

adj_matrix <- matrix(c(0, 1, 0, 1, 0,
                       1, 0, 1, 0, 0,
                       0, 1, 0, 0, 0,
                       1, 0, 0, 0, 1,
                       0, 0, 0, 1, 0), ncol = 5)

edges <- create_edges_visnetwork(adj_matrix)

shared_edges <- data.frame(from = c(1,2), to = c(4, 3))

edges <- set_attributes_to_edges_visnetwork(edges,
                                            subset_edges = shared_edges,
                                            width = c(3, .5),
                                            color = c("red", "blue"))
\dontrun{                                                                                    
if(requireNamespace("visnetwork", quietly = TRUE) & interactive()){
  visnetwork(nodes, edges) 
} else {
  print("install the 'igraph' package")
}
}
}
