% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-graph.R
\name{create_nodes_visnetwork}
\alias{create_nodes_visnetwork}
\title{Nodes for the \code{visNetwork} package}
\usage{
create_nodes_visnetwork(n_nodes, labels = 1:n_nodes)
}
\arguments{
\item{n_nodes}{Number of nodes in the graph}

\item{labels}{The labels for the individual nodes
(Default: \code{1:n_nodes})}
}
\value{
Data frame with two columns: \code{id} and \code{title}
}
\description{
Creates a data frame that can be used for the
\code{visNetwork} package.
}
\examples{
nodes <- create_nodes_visnetwork(n_nodes = 5, labels = LETTERS[1:5])

adj_matrix <- matrix(c(0, 1, 0, 1, 0,
                       1, 0, 1, 0, 0,
                       0, 1, 0, 0, 0,
                       1, 0, 0, 0, 1,
                       0, 0, 0, 1, 0), ncol = 5)

edges <- create_edges_visnetwork(adj_matrix)

shared_edges <- data.frame(from = c(1,2), to = c(4, 3))

edges <- set_attributes_to_edges_visnetwork(edges,
                                            subset_edges = shared_edges,
                                            width = c(3, .5),
                                            color = c("red", "blue"))

visnetwork(nodes, edges) 
}
