% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-weight-matrices.R
\name{create_weight_matrix}
\alias{create_weight_matrix}
\title{Different Weight Matrices}
\usage{
create_weight_matrix(
  type = c("full", "glasso", "grid", "uniform-random"),
  m = 9
)
}
\arguments{
\item{type}{The type of weight matrix}

\item{m}{Number of graphs (works currently only for m = 9)}
}
\value{
Weight matrix
}
\description{
This function generates different weight matrices, \eqn{W}. 
There are several types: 
\describe{
\item{\code{full} }{All graphs are fully connected with weight 1}
\item{\code{glasso} }{All graphs are disconnected with weight 0. This 
             mimicks the GLASSO, where each graph is estimated independently}
\item{\code{grid} }{A weight matrix for a \eqn{3 \times 3} grid}
\item{\code{uniform-random} }{Fully-connected, but the entries are drawn 
                   from a uniform distribution}
}
}
