% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{CVN-package}
\alias{CVN-package}
\title{CVN}
\description{
An estimator for graphical models changing with multiple discrete
external covariates
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bips-hb/CVN}
  \item \url{https://arxiv.org/abs/2407.19978}
  \item Report bugs at \url{https://github.com/bips-hb/CVN/issues}
}

}
\author{
Louis Dijkstra
}
\keyword{internal}
