% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVN-package.R
\docType{package}
\name{CVN-package}
\alias{CVN-package}
\title{Covariate-varying Networks}
\description{
Inferring high-dimensional Gaussian graphical networks that
change with multiple discrete covariates.
}
\examples{

data(grid)
W <- create_weight_matrix(type = "grid", k=3, l=3, plot = FALSE)

cvn <- CVN(grid, W, lambda1 = 1, lambda2 = 1:2, 
           n_cores = 1,
           eps = 1e-2, maxiter = 1000, verbose = TRUE)
}
\references{
Dijkstra L, Godt A, Foraita R
\emph{Inferring High-Dimensional Dynamic Networks Changing with Multiple Covariates (2024), Arxiv},
\url{https://arxiv.org/abs/2407.19978}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bips-hb/CVN}
  \item \url{https://bips-hb.github.io/CVN/}
  \item Report bugs at \url{https://github.com/bips-hb/CVN/issues}
}

}
\author{
\strong{Maintainer}: Ronja Foraita \email{foraita@leibniz-bips.de} (\href{https://orcid.org/0000-0003-2216-6653}{ORCID})

Authors:
\itemize{
  \item Louis Dijkstra (\href{https://orcid.org/0000-0003-1476-6202}{ORCID})
}

Other contributors:
\itemize{
  \item Lukas Burk (\href{https://orcid.org/0000-0001-7528-3795}{ORCID}) [contributor]
  \item Leibniz Institute for Prevention Research and Epidemiology - BIPS \email{R@leibniz-bips.de} (02c22vc57) [copyright holder]
}

}
\keyword{graphical}
\keyword{models}
