% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVN.R
\name{CVN}
\alias{CVN}
\title{Inferring a Covariate-Varying Network (CVN)}
\usage{
CVN(
  data,
  W,
  lambda1 = 1,
  lambda2 = 1,
  rho = 1,
  epsilon = 10^(-5),
  maxiter = 100,
  n_cores = 1,
  normalized = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{data}{A list with matrices. The number of columns should be the 
same for each matrix. Number of observations can differ}

\item{W}{The \eqn{(m x m)}-dimensional uppertriangular weight matrix}

\item{lambda1}{The \eqn{\lambda1} LASSO penalty term}

\item{lambda2}{The \eqn{\lambda2} global smoothing parameter}

\item{rho}{The \eqn{\rho} ADMM's penalty parameter (Default: 1)}

\item{epsilon}{If the relative difference between two update steps is 
smaller than \eqn{\epsilon}, the algorithm stops
(Default: 10^-5)}

\item{maxiter}{Maximum number of iterations (Default = 100)}

\item{n_cores}{Number of cores used (Default: 1)}

\item{normalized}{Data is normalized if TRUE. Otherwise the data is only
centered (Default: FALSE)}

\item{verbose}{Verbose (Default: FALSE)}
}
\value{
A CVN data object; a list with entries
  \item{\code{data}}{The \code{raw_data}, but then normalized or centered}
  \item{\code{W}}{The \eqn{m x m} weight matrix}
  \item{\code{m}}{Number of graphs}
  \item{\code{p}}{Number of variables}
  \item{\code{n_obs}}{Number of observations for each graph}
}
\description{
Estimates a covariate-varying network model (CVN), i.e., \eqn{m} 
Gaussian graphical models that change with (multiple) external covariate(s). 
The smoothing between the graphs is specified by the \eqn{(m x m)}-dimensional
weight matrix \eqn{W}. The function returns the estimated precision matrices 
for each graph.
}
