% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVN.R
\name{CVN}
\alias{CVN}
\title{Estimating a Covariate-Varying Network (CVN)}
\usage{
CVN(
  data,
  W,
  lambda1 = 1,
  lambda2 = 1,
  rho = 1,
  epsilon = 10^(-5),
  maxiter = 100,
  n_cores = 1,
  truncate = 1e-05,
  normalized = FALSE,
  warmstart = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{data}{A list with matrices. The number of columns should be the 
same for each matrix. Number of observations can differ}

\item{W}{The \eqn{(m \times m)}-dimensional upper-triangular 
weight matrix \eqn{W}}

\item{lambda1}{The \eqn{\lambda_1} LASSO penalty term}

\item{lambda2}{The \eqn{\lambda_2} global smoothing parameter}

\item{rho}{The \eqn{\rho} ADMM's penalty parameter (Default: \code{1})}

\item{epsilon}{If the relative difference between two update steps is 
smaller than \eqn{\epsilon}, the algorithm stops. 
See \code{\link{relative_difference_precision_matrices}}
(Default: \code{1e-5})}

\item{maxiter}{Maximum number of iterations (Default: \code{100})}

\item{n_cores}{Number of cores used (Default: \code{1})}

\item{truncate}{All values of the final \eqn{\hat{\Theta}_i}'s below \code{truncate} will be 
set to \code{0}. (Default: \code{1e-5})}

\item{normalized}{Data is normalized if \code{TRUE}. Otherwise the data is only
centered (Default: \code{FALSE})}

\item{warmstart}{If \code{TRUE}, use the \code{\link[huge]{huge}} package for estimating
the individual graphs first (Default: \code{FALSE})}

\item{verbose}{Verbose (Default: \code{FALSE})}
}
\value{
A \code{CVN} object; a list with entries
   \item{\code{Theta}}{The estimated precision matrices \eqn{\{ \hat{\Theta}_i \}_{i = 1}^m}}
   \item{\code{adj_matrices}}{The estimated adjacency matrices; 
                        \code{1} if there is an edge, \code{0} otherwise. 
                        Matrices are sparse using package \code{\link[Matrix]{Matrix}}}
   \item{\code{Sigma}}{Empirical covariance matrices \eqn{\{\hat{\Sigma}_i\}_{i = 1}^m}}
   \item{\code{m}}{Number of graphs}
   \item{\code{p}}{Number of variables}
   \item{\code{n_obs}}{Vector of length \eqn{m} with number of observations for each graph}
  \item{\code{data}}{The \code{data}, but then normalized or centered}
  \item{\code{normalized}}{If \code{TRUE}, \code{data} was normalized. Otherwise \code{data} was only centered}
  \item{\code{W}}{The \eqn{(m \times m)}-dimensional weight matrix \eqn{W}}
  \item{\code{D}}{Matrix \eqn{D} used for the Generalized LASSO}
  \item{\code{lambda1}}{The \eqn{\lambda_1} LASSO penalty term}
  \item{\code{lambda2}}{The \eqn{\lambda_2} global smoothing parameter} 
  \item{\code{rho}}{The \eqn{\rho} ADMM's penalty parameter} 
  \item{\code{epsilon}}{The stopping criterion \eqn{\epsilon}} 
  \item{\code{converged}}{If \code{TRUE}, stopping condition has been met; \code{FALSE} otherwise}
  \item{\code{value}}{The final relative difference}
  \item{\code{n_iterations}}{Number of iterations}
  \item{\code{truncate}}{Truncation value for \eqn{\{ \hat{\Theta}_i \}_{i = 1}^m}}
}
\description{
Estimates a covariate-varying network model (CVN), i.e., \eqn{m} 
Gaussian graphical models that change with (multiple) external covariate(s). 
The smoothing between the graphs is specified by the \eqn{(m \times m)}-dimensional
weight matrix \eqn{W}. The function returns the estimated precision matrices 
for each graph.
}
