% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamming-distance.R
\name{hamming_distance_adj_matrices}
\alias{hamming_distance_adj_matrices}
\title{Structural Hamming Distance}
\usage{
hamming_distance_adj_matrices(adj_matrices)
}
\arguments{
\item{adj_matrices}{A list of adjacency matrices}
}
\value{
Matrix of Hamming distances
}
\description{
Returns the structural Hamming distance between multiple
graphs
}
\examples{
path <- system.file("cvnfit.rda", package = "CVN")
load(path)

hamming_distance_adj_matrices(fit$adj_matrices[[1]])

}
