% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.cvn.R
\name{plot.cvn}
\alias{plot.cvn}
\title{Plot Function for CVN Object Class}
\usage{
\method{plot}{cvn}(x, ...)
}
\arguments{
\item{x}{\code{cvn} object}

\item{...}{Additional arguments to pass to \code{\link{visnetwork_cvn}}}
}
\description{
Custom plot method for CVN objects.
}
\examples{
\donttest{
path <- system.file("cvnfit.rda", package = "CVN")
load(path)
fit <- plot(fit) 
fit$plots[[1]][[4]]
}
}
\seealso{
\code{CVN}
}
