% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_weight_matrix.R
\name{plot_weight_matrix}
\alias{plot_weight_matrix}
\title{Plot weight matrix as a grid}
\usage{
plot_weight_matrix(W, k, l)
}
\arguments{
\item{W}{The weight matrix which equals an adjacency matrix}

\item{k}{Number of categories in the first external covariate}

\item{l}{Number of categories in the second external covariate}
}
\value{
An igraph object
}
\description{
This function requires the igraph package to be installed.
}
\examples{
# Requires to have the igraph package to be installed
W <- create_weight_matrix(type = "full", k = 2, l = 3)
plot_weight_matrix(W, k=2, l=3)

}
\seealso{
\code{\link{hd_weight_matrix}}
}
