% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.cvn}
\alias{print.cvn}
\title{Print CVN}
\usage{
\method{print}{cvn}(x, ...)
}
\arguments{
\item{x}{Object of classs 'CVN'}

\item{...}{Additional arguments to pass}
}
\description{
Print contents of CVN object
}
\seealso{
\code{\link{CVN}}
}
