% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-graph.R
\name{set_attributes_to_edges_visnetwork}
\alias{set_attributes_to_edges_visnetwork}
\title{Add Attributes to Subset of Edges for \code{visNetwork}}
\usage{
set_attributes_to_edges_visnetwork(
  edges,
  subset_edges,
  width = c(NA, NA),
  color = c(NULL, NULL)
)
}
\arguments{
\item{edges}{A data.frame create by \code{\link{create_edges_visnetwork}}}

\item{subset_edges}{A list with the elements \code{from} and \code{to}. Both
\code{from} and \code{to} are vectors of the same length
denoting the different edges}

\item{width}{Vector with two values. The first is assigned to the
edges in the subset given by \code{subset_edges}. The second
value is assigned to the rest. If \code{width = c(NA,NA)},
no width is assigned}

\item{color}{Vector with two values. The first is assigned to the
edges in the subset given by \code{subset_edges}. The second
value is assigned to the rest. If \code{color = c(NULL,NULL)},
no color is assigned}
}
\value{
A data frame that can be used by the \code{visNetwork} package
}
\description{
A subset of edges can be assign a different thickness
or color.
}
