% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-Z.R
\name{updateZ}
\alias{updateZ}
\title{The \eqn{Z}-update step for the ADMM}
\usage{
updateZ(
  m,
  p,
  nrow_D,
  Theta,
  Y,
  W,
  eta1,
  eta2,
  a,
  rho_genlasso,
  maxiter_genlasso,
  eps_genlasso,
  truncate_genlasso,
  use_genlasso_package
)
}
\arguments{
\item{m}{Number of graphs}

\item{p}{Number of variables}

\item{Theta}{A list with matrices with the current values of \eqn{\Theta}}

\item{Y}{A list with matrices with the current values of \eqn{Y}}

\item{D}{A matrix used for solving the Generalized LASSO}
}
\value{
A list with matrices with the new values of \eqn{Z}
}
\description{
Returns the updated value of \eqn{Z} for the ADMM given the previously 
updated values of \eqn{\Theta} and \eqn{Y}
}
\seealso{
\code{\link{create_matrix_D}}
}
