% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-Z-wrapper.R
\name{updateZ_wrapper}
\alias{updateZ_wrapper}
\title{Wrapper for the \eqn{Z}-update Step for the ADMM}
\usage{
updateZ_wrapper(
  m,
  p,
  nrow_D,
  Theta,
  Y,
  W,
  eta1,
  eta2,
  a,
  rho_genlasso,
  maxiter_genlasso,
  eps_genlasso,
  truncate_genlasso
)
}
\arguments{
\item{m}{Number of graphs}

\item{p}{Number of variables}

\item{nrow_D}{Number of rows of the \eqn{D}-matrix}

\item{Theta}{A list with matrices with the current values of \eqn{\Theta}}

\item{Y}{A list with matrices with the current values of \eqn{Y}}

\item{D}{A matrix used for solving the Generalized LASSO}
}
\value{
A list with matrices with the new values of \eqn{Z}
}
\description{
A wrapper for the \code{C} function that returns the 
updated value of \eqn{Z} for the ADMM given the previously 
updated values of \eqn{\Theta} and \eqn{Y}
}
\seealso{
\code{\link{create_matrix_D}}
}
