% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{updateZRcpp}
\alias{updateZRcpp}
\title{The \eqn{Z}-update Step
A \code{C} implementation of the \eqn{Z}-update step. We
solve a generalized LASSO problem repeatedly for each of the
individual edges}
\usage{
updateZRcpp(m, p, Theta, Y, W, eta1, eta2, a, rho, max_iter, eps, truncate)
}
\arguments{
\item{m}{The number of graphs}

\item{p}{The number of variables}

\item{Theta}{A list of matrices with the \eqn{\Theta}-matrices}

\item{Y}{A list of matrices with the \eqn{Y}-matrices}

\item{W}{The weight matrix \eqn{W} of dimensions \eqn{m x m}}

\item{eta1}{Equals \eqn{\lambda_1 / rho}}

\item{eta2}{Equals \eqn{\lambda_2 / rho}}

\item{a}{Value added to the diagonal of \eqn{-D'D} so that
the matrix is positive definite, see
\code{\link{matrix_A_inner_ADMM}}}

\item{rho}{The ADMM's parameter}

\item{max_iter}{Maximum number of iterations}

\item{eps}{Stopping criterion. If differences
are smaller than \eqn{\epsilon}, algorithm
is halted}

\item{truncate}{Values below \code{truncate} are
set to \code{0}}
}
\value{
The estimated vector \eqn{\hat{\beta}}
}
\description{
The \eqn{Z}-update Step
A \code{C} implementation of the \eqn{Z}-update step. We
solve a generalized LASSO problem repeatedly for each of the
individual edges
}
