% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-graph.R
\name{visnetwork_cvn}
\alias{visnetwork_cvn}
\title{All \code{visNetwork} plots for a CVN object}
\usage{
visnetwork_cvn(
  cvn,
  node_titles = 1:cvn$p,
  titles = lapply(1:cvn$n_lambda_values, function(i) sapply(1:cvn$m, function(j) "")),
  show_core_graph = TRUE,
  width = c(3, 1),
  color = c("red", "blue"),
  igraph_layout = "layout_in_circle",
  verbose = TRUE
)
}
\arguments{
\item{cvn}{A \code{cvn} object, see \code{\link{CVN}}}

\item{node_titles}{Vector with title of the nodes (Default: \code{1:p})}

\item{titles}{A list with \code{n_lambda_values} vectors. Each vector is of the
length \code{m}. Regulates the titles of the graphs (Default: no title)}

\item{show_core_graph}{Shall the core graph be visualized (Default = TRUE)}

\item{width}{Edge width of the core graph}

\item{color}{String vector with two colors. The first color marks the edges in the
core graph (Default: c("red", "blue"))}

\item{igraph_layout}{igraph layout (default: layout_in_circle)}

\item{verbose}{Verbose (Default: \code{TRUE})}
}
\value{
The cvn input which is extended by the list element 'plots'
}
\description{
Creates all \code{visNetwork} plots, see \code{\link{visnetwork}},
for all graphs in a \code{cvn} object
}
\examples{
\dontrun{
path <- system.file("cvnfit.rda", package = "CVN")
load(path)
fit_plot <- visnetwork_cvn(fit)
fit_plot$plots[[1]][[1]]
}
}
\seealso{
\code{\link{CVN}}, \code{\link{visnetwork}}
}
