% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-graph.R
\name{visnetwork_cvn}
\alias{visnetwork_cvn}
\title{All \code{visNetwork} plots for a CVN object}
\usage{
visnetwork_cvn(
  cvn,
  node_titles = 1:cvn$p,
  titles = lapply(1:cvn$n_lambda_values, function(i) sapply(1:cvn$m, function(j) "")),
  show_core_graph = TRUE,
  width = c(3, 1),
  color = c("red", "blue"),
  igraph_layout = "layout_in_circle",
  verbose = TRUE
)
}
\arguments{
\item{cvn}{A \code{cvn} object, see \code{\link{CVN::CVN}} 
or \code{\link{CVN::glasso}}}

\item{node_titles}{Vector with title of the nodes (Default: \code{1:p})}

\item{titles}{A list with \code{n_lambda_values} vectors. Each vector is of the 
lenght \code{m}. Regulates the titles of the graphs (Default: no title)}

\item{show_core_graph, width, color}{Show the core graph using the width and colors}
}
\value{
List
}
\description{
Creates all \code{visNetwork} plots, see \code{\link{CVN::visnetwork}}, 
for all graphs in a \code{cvn} object
}
